requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-allauth
django-rest-auth
python-decouple
psycopg2-binary
pytest
pytest-django
pytest-cov
django-cors-headers
django-storages
boto3
pypdf
celery
redis
